+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/Watchable.java
@@ -15,10 +15,36 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
+import io.fabric8.kubernetes.api.model.ListOptions;
+
 public interface Watchable<H, W> {
 
+  /**
+   * Watch returns {@link H} interface that watches requested resource
+   *
+   * @param watcher Watcher interface of Kubernetes resource
+   * @return watch interface {@link H}
+   */
     H watch(W watcher);
 
+  /**
+   * Watch returns {@link H} interface that watches requested resource
+   *
+   * @param options options available for watch operation
+   * @param watcher Watcher interface of Kubernetes resource
+   * @return watch interface {@link H}
+   */
+    H watch(ListOptions options, W watcher);
+
+  /**
+   * Watch returns {@link H} interface that watches requested resource from
+   * specified resourceVersion
+   *
+   * @param resourceVersion resource version from where to start watch
+   * @param watcher Watcher interface of Kubernetes resource
+   * @deprecated Please use {@link #watch(ListOptions, Object)} instead, it has a parameter of resourceVersion
+   * @return watch interface {@link H}
+   */
     @Deprecated
     H watch(String resourceVersion, W watcher);